chore(Wallet): refactor error codes
