--- data/abi/OmnuumWalletAdv.json
@@ -1,26 +1,0 @@
-[
-  "constructor(uint256,uint8,tuple(address,uint8)[])",
-  "event Approved(address indexed,uint256 indexed,uint8)",
-  "event Canceled(address indexed,uint256 indexed)",
-  "event EtherReceived(address indexed)",
-  "event Executed(address indexed,uint256 indexed,uint8 indexed)",
-  "event PaymentReceived(address indexed,bytes32 indexed,string)",
-  "event Requested(address indexed,uint256 indexed,uint8 indexed)",
-  "event Revoked(address indexed,uint256 indexed,uint8)",
-  "function approve(uint256)",
-  "function cancel(uint256)",
-  "function consensusRatio() view returns (uint256)",
-  "function execute(uint256)",
-  "function getLastRequestNo() view returns (uint256)",
-  "function isOwner(address) view returns (bool)",
-  "function isOwnerVoted(address,uint256) view returns (bool)",
-  "function makePayment(bytes32,string) payable",
-  "function minLimitForConsensus() view returns (uint8)",
-  "function ownerCounter(uint8) view returns (uint8)",
-  "function ownerVote(address) view returns (uint8)",
-  "function request(uint8,tuple(address,uint8),tuple(address,uint8),uint256)",
-  "function requests(uint256) view returns (address, uint8, tuple(address,uint8), tuple(address,uint8), uint256, uint256, bool)",
-  "function requiredVotesForConsensus() view returns (uint256)",
-  "function revoke(uint256)",
-  "function totalVotes() view returns (uint256)"
-]
