+++ utils/constants.js
@@ -61,24 +61,34 @@ module.exports = {
     RevertMessage: {
       silent: 'Transaction reverted silently',
     },
-    wallet: {
-      useless: 'Useless payment',
-      onlyOwner: 'only owner',
-      reqNotExists: 'request not exist',
-      alreadyApproved: 'already approved',
-      notApproved: 'not approved',
-      alreadyWithdrawn: 'already withdrawn',
-      consensusNotReached: 'consensus not reached',
-      notEnoughBalance: 'request value exceeds balance',
-      notRequester: 'withdrawer must be the requester',
-    },
     code: {
-      ARG1: 'ARG1',
-      ARG2: 'ARG2',
-      OO1: 'OO1',
-      OO2: 'OO2',
-      OO3: 'OO3',
+      // Arguments Error
+      ARG1: 'ARG1', // Arguments length should be same
+      ARG2: 'ARG2', // Arguments are not correct
+      // Only Owner Error
+      OO1: 'OO1', // Ownable: Caller is not the collection owner
+      OO2: 'OO2', // Only Omnuum or owner can change
+      OO3: 'OO3', // Only Omnuum can call
+      OO4: 'OO4', // Only the owner of the wallet is allowed
+      OO5: 'OO5', // Already the owner of the wallet
+      OO6: 'OO6', // Only the requester is allowed
+      // Not Exist Error
+      NX1: 'NX1', // ERC721Metadata: URI query for nonexistent token
+      NX2: 'NX2', // Non-existent wallet account
+      NX3: 'NX3', // Non-existent owner request
+      // Number Error
       NE1: 'NE1', // Fee rate should be lower than 100%
+      NE2: 'NE2', // Not reach consensus
+      NE3: 'NE3', // A zero payment is not acceptable
+      NE4: 'NE4', // Insufficient balance
+      NE5: 'NE5', // Violate min limit for consensus
+      // State Error
+      SE1: 'SE1', // Already executed
+      SE2: 'SE2', // Request canceled
+      SE3: 'SE3', // Already voted
+      SE4: 'SE4', // Not voted
+      SE5: 'SE5', // Address: unable to send value, recipient may have reverted
+      // Mint Error
       MT1: 'MT1', // There is no available ticket
       MT2: 'MT2', // Cannot mint more than possible amount per address
       MT3: 'MT3', // Remaining token count is not enough
@@ -86,6 +96,9 @@ module.exports = {
       MT7: 'MT7', // Mint is ended
       MT8: 'MT8', // Minting period is ended
       MT9: 'MT9', // Mint subject cannot be CA
+      // Address Error
+      AE1: 'AE1', // Zero address not acceptable
+      AE2: 'AE2', // Contract address not acceptable
     },
   },
   payloadTopic: {
